[#50][preview] setup submit handlers for the ChatPage
