+++ src/pages/ChatPage/ChatPage.html
@@ -57,8 +57,8 @@
     <ypr-profile-input name="avatar"></ypr-profile-input>
     <ypr-form>
       <ypr-input
+        slot="content"
         label="Email"
-        value="email@mail.com"
         details="User email"
         type="email"
         name="emai"
@@ -67,8 +67,8 @@
         info
       ></ypr-input>
       <ypr-input
+        slot="content"
         label="Username"
-        value="username"
         details="User name"
         name="login"
         validate="isLogin"
@@ -76,8 +76,8 @@
         info
       ></ypr-input>
       <ypr-input
+        slot="content"
         label="First Name"
-        value="John"
         name="first_name"
         details="First name user"
         validate="isDisplayName"
@@ -85,8 +85,8 @@
         info
       ></ypr-input>
       <ypr-input
+        slot="content"
         label="Last Name"
-        value="Doe"
         name="second_name"
         details="Last name user"
         validate="isDisplayName"
@@ -94,8 +94,8 @@
         info
       ></ypr-input>
       <ypr-input
+        slot="content"
         label="Display Name"
-        value="John Doe"
         name="display_name"
         details="Name that displayed"
         validate="isDisplayName"
@@ -104,8 +104,8 @@
       >
       </ypr-input>
       <ypr-input
+        slot="content"
         label="Phone Number"
-        value="+7 (000) 000-00-00"
         name="phone"
         details="Phone number user"
         type="phone"
@@ -123,12 +123,18 @@
         validateErrorMessage="Password is not correct"
         info
       ></ypr-input>
-      <ypr-input label="Confirm Password" name="password" details="Password field" info></ypr-input>
+      <ypr-input
+        slot="content"
+        label="Confirm Password"
+        name="password"
+        details="Password field"
+        info
+      ></ypr-input>
     </ypr-form>
   </ypr-main>
 
   <ypr-footer slot="settings-footer">
-    <ypr-btn>
+    <ypr-btn id="btnSubmit">
       Save
       <ypr-icon name="done" color="var(--gray-400)"></ypr-icon>
     </ypr-btn>
